[client] [dashboard] [accounts_head] [new_account_form] :: Isolating New_Account_Form as new component
