+++ client/src/components/AccountsManagement/AccountsHead/AccountsHead.jsx
@@ -1,14 +1,12 @@
-import { AppstoreAddOutlined, BankOutlined, CheckOutlined, HomeOutlined, MoneyCollectTwoTone, UserOutlined } from '@ant-design/icons';
-import { Button, Form, Input, Modal, Select, Table } from 'antd';
+import { AppstoreAddOutlined, CheckOutlined, HomeOutlined, MoneyCollectTwoTone, UserOutlined } from '@ant-design/icons';
+import { Button, Modal, Table } from 'antd';
 import React, { useEffect, useState } from 'react';
-import config from '../../../config/config';
 import NavigationPath from '../../NavigationPath/NavigationPath';
 import TableTitle from '../../TableTitle/TableTitle';
 import './AccountsHead.css';
 import Columns from './models/TableColumns';
+import NewAccountHeadForm from './NewAccountHeadForm';
 import LoadTableData from './xhr/getTableData';
-
-const { Option } = Select;
 
 const navigationPath = [
   {
@@ -62,31 +60,6 @@ const AccountsHead = () => {
     /* Empty array means, this hook is run only once when the component is mounted */
   }, []);
 
-  const imitateNumberInput = (e) => {
-    const { value } = e.target;
-    if (/^-?\d*(\.\d*)?$/.test(value) || value === '' || value === '-') {
-      formRef.current.setFieldsValue({
-        account_opening_balance: value,
-      });
-    } else {
-      formRef.current.setFieldsValue({
-        account_opening_balance: value.slice(0, -1),
-      });
-    }
-  };
-
-  const formatValue = (e) => {
-    let { value } = e.target;
-
-    if (value.charAt(value.length - 1) === '.' || value === '-') {
-      value = value.slice(0, -1);
-    }
-
-    formRef.current.setFieldsValue({
-      account_opening_balance: parseFloat(value).toFixed(2),
-    });
-  };
-
   const submitNewAccountForm = async (values) => {
     console.log('form submitted');
     console.log(values);
@@ -139,83 +112,7 @@ const AccountsHead = () => {
           </Button>,
         ]}
       >
-
-        <Form
-          name="new-account-form"
-          ref={formRef}
-          onFinish={submitNewAccountForm}
-          hideRequiredMark
-          layout="vertical"
-        >
-
-          <Form.Item
-            name="account_type"
-            label="Account Type"
-            rules={[
-              {
-                required: true,
-                message: 'Account Type Must Be Selected!',
-              },
-            ]}
-          >
-            <Select
-              showSearch
-              size="large"
-              placeholder="Account Type"
-              optionFilterProp="children"
-              filterOption={(input, option) => option.children.toLowerCase().indexOf(input.toLowerCase()) >= 0}
-              disabled={state.newAccountHeadLoading}
-            >
-              <Option value="bank">Bank Account</Option>
-              <Option value="expense">Expense Account</Option>
-            </Select>
-          </Form.Item>
-
-          <Form.Item
-            name="account_head_name"
-            label="Account Head Name"
-            rules={[
-              {
-                required: true,
-                message: 'Account Head Name Cannot Be Empty!',
-              },
-            ]}
-          >
-            <Input
-              size="large"
-              placeholder="Account Head Name"
-              prefix={<BankOutlined />}
-              disabled={state.newAccountHeadLoading}
-              onPressEnter={() => formRef.current.submit()}
-            />
-          </Form.Item>
-
-          <Form.Item
-            name="account_opening_balance"
-            label="Opening Balance (On Account Creation)"
-            initialValue="0.00"
-            rules={[
-              {
-                required: true,
-                message: 'Account Opening Balance Cannot Be Empty!',
-              },
-            ]}
-          >
-            <Input
-              addonBefore={config.LOCALE.currencySymbol}
-              disabled={state.newAccountHeadLoading}
-              size="large"
-              style={{
-                width: '100%',
-              }}
-              onPressEnter={() => formRef.current.submit()}
-              onChange={imitateNumberInput}
-              onBlur={formatValue}
-            />
-          </Form.Item>
-
-        </Form>
-
+        <NewAccountHeadForm onSubmit={submitNewAccountForm} state={state} formRef={formRef} />
       </Modal>
     </div>
   );
